fix: add unit test case, use isSPFXProject function in core
